requirements.txt
# AI Data Agent - Complete Requirements

# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6

# Redis
redis==5.0.1
aioredis==2.0.1

# AI and ML
groq==0.4.1
openai==1.6.1
langchain==0.1.0
langchain-community==0.0.13
numpy==1.24.3
pandas==2.0.3

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# File Processing
python-magic==0.4.27
filetype==1.2.0
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
loguru==0.7.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# CORS and Middleware
starlette==0.27.0

# Date and Time
python-dateutil==2.8.2

# Environment and Configuration
pyyaml==6.0.1

# Monitoring and Logging
structlog==23.2.0

# Type Hints
typing-extensions==4.8.0

# Production Server
gunicorn==21.2.0